Moved prisma client invokation to root dir
