+++ backend/src/routes/prompt.ts
@@ -1,5 +1,5 @@
 import express from "express";
-import { prisma } from "../../prisma/prisma";
+import { prisma } from "../prisma";
 import { getResponseForPrompt } from "./gemini";
 
 export const promptRouter = express.Router();
